# Add timing data to `valout`