--- a/examples/multi-layer/bowl-radial/Makefile
+++ b/examples/multi-layer/bowl-radial/Makefile
@@ -21,43 +21,14 @@ PLOTDIR = _plots               # Directory for plots
 # Environment variable FC should be set to fortran compiler, e.g. gfortran
 
 # Compiler flags can be specified here or set as an environment variable
-FFLAGS ?= 
-
-# Include flags for LAPACK linking
-UNAME = $(shell uname)
-ifeq ($(UNAME), Darwin)
-	ifndef LFLAGS
-		LFLAGS = -framework accelerate $(FFLAGS)
-	else
-		LFLAGS += -framework accelerate $(FFLAGS)
-	endif
-else ifeq ($(UNAME), Linux)
-	ifeq ($(FC), ifort)
-		# Created using Intel MKL Link Line Advisor
-		# Please refer to 
-		# http://software.intel.com/en-us/articles/intel-mkl-link-line-advisor/
-		# for other installation types
-		FFLAGS += -I$(MKLROOT)/include
-		ifndef LFLAGS                                                                                                                                    
-			LFLAGS =  -Wl,--start-group $(MKLROOT)/lib/intel64/libmkl_intel_lp64.a $(MKLROOT)/lib/intel64/libmkl_core.a $(MKLROOT)/lib/intel64/libmkl_sequential.a -Wl,--end-group -lpthread -lm $(FFLAGS)
-		else
-			LFLAGS +=  -Wl,--start-group $(MKLROOT)/lib/intel64/libmkl_intel_ilp64.a $(MKLROOT)/lib/intel64/libmkl_core.a $(MKLROOT)/lib/intel64/libmkl_sequential.a -Wl,--end-group -lpthread -lm $(FFLAGS)
-		endif
-	else ifeq ($(FC), gfortran)
-		ifndef LFLAGS
-			LFLAGS = -llapack $(FFLAGS)
-		else
-			LFLAGS += -llapack $(FFLAGS)
-		endif
-	endif
-endif
+FFLAGS ?=
 
 # ---------------------------------
 # package sources for this program:
 # ---------------------------------
 
 GEOLIB = $(CLAW)/geoclaw/src/2d/shallow
-include $(GEOLIB)/Makefile.geoclaw
+include $(GEOLIB)/multilayer/Makefile.multilayer
 
 # ---------------------------------------
 # package sources specifically to exclude
@@ -74,29 +45,13 @@ EXCLUDE_SOURCES = \
 # ----------------------------------------
 
 MODULES = \
- $(GEOLIB)/multilayer/gauges_module.f90 \
  ./qinit_module.f90
 
 SOURCES = \
   ./qinit.f90 \
-  $(GEOLIB)/multilayer/setaux.f90 \
-  $(GEOLIB)/multilayer/b4step2.f90 \
-  $(GEOLIB)/multilayer/getmaxspeed.f90 \
-  $(GEOLIB)/multilayer/stepgrid.f \
-  $(GEOLIB)/multilayer/valout.f90 \
-  $(GEOLIB)/multilayer/src2.f90 \
-  $(GEOLIB)/multilayer/filpatch.f90 \
-  $(GEOLIB)/multilayer/filval.f90 \
-  $(GEOLIB)/multilayer/flag2refine2.f90 \
-  $(GEOLIB)/multilayer/step2.f90 \
-  $(GEOLIB)/multilayer/upbnd.f \
-  $(GEOLIB)/multilayer/update.f90 \
   $(CLAW)/riemann/src/rpn2_layered_shallow_water.f90 \
   $(CLAW)/riemann/src/rpt2_layered_shallow_water.f90 \
   $(CLAW)/riemann/src/geoclaw_riemann_utils.f
-  # $(CLAW)/riemann/src/rpn2_geoclaw.f \
-  # $(CLAW)/riemann/src/rpt2_geoclaw.f \
-  # $(CLAW)/riemann/src/geoclaw_riemann_utils.f \
 
 #-------------------------------------------------------------------
 # Include Makefile containing standard definitions and make options: