--- a/src/2d/shallow/amr2.f90
+++ b/src/2d/shallow/amr2.f90
@@ -117,6 +117,13 @@ program amr2
     ! Timing variables
     integer :: clock_start, clock_finish, clock_rate, ttotal
     real(kind=8) :: ttotalcpu
+    character(len=256) :: timing_line, timing_substr
+    character(len=*), parameter :: timing_file_name = "timing.csv"
+    integer, parameter :: out_unit = 67
+    character(len=*), parameter :: timing_header_format =                      &
+                                                  "(' wall time (', i2,')," // &
+                                                  " CPU time (', i2,'), "   // &
+                                                  "cells updated (', i2,'),')"
 
     ! Common block variables
     real(kind=8) :: dxmin, dymin
@@ -475,6 +482,19 @@ program amr2
         call set_gauges(rest, nvar, naux) ! Set gauge output
 
     else
+
+        ! Create new timing file
+        open(unit=out_unit, file=timing_file_name, form='formatted',         &
+             status='unknown', action='write')
+        ! Construct header string
+        timing_line = 'output_time,total_wall_time,total_cpu_time,'
+        timing_substr = ""
+        do level=1, mxnest
+            write(timing_substr, timing_header_format) level, level, level
+            timing_line = trim(timing_line) // trim(timing_substr)
+        end do
+        write(out_unit, "(a)") timing_line
+        close(out_unit)
 
         open(outunit, file=outfile, status='unknown', form='formatted')