--- a/src/2d/shallow/multilayer/Makefile.multilayer
+++ b/src/2d/shallow/multilayer/Makefile.multilayer
@@ -1,0 +1,184 @@
+# Include flags for LAPACK linking
+UNAME = $(shell uname)
+ifeq ($(UNAME), Darwin)
+  ifndef LFLAGS
+    LFLAGS = -framework accelerate $(FFLAGS)
+  else
+    LFLAGS += -framework accelerate $(FFLAGS)
+  endif
+else ifeq ($(UNAME), Linux)
+  ifeq ($(FC), ifort)
+    # Created using Intel MKL Link Line Advisor
+    # Please refer to 
+    # http://software.intel.com/en-us/articles/intel-mkl-link-line-advisor/
+    # for other installation types
+    FFLAGS += -I$(MKLROOT)/include
+    ifndef LFLAGS                                                                                                                                    
+      LFLAGS =  -Wl,--start-group $(MKLROOT)/lib/intel64/libmkl_intel_lp64.a $(MKLROOT)/lib/intel64/libmkl_core.a $(MKLROOT)/lib/intel64/libmkl_sequential.a -Wl,--end-group -lpthread -lm $(FFLAGS)
+    else
+      LFLAGS +=  -Wl,--start-group $(MKLROOT)/lib/intel64/libmkl_intel_ilp64.a $(MKLROOT)/lib/intel64/libmkl_core.a $(MKLROOT)/lib/intel64/libmkl_sequential.a -Wl,--end-group -lpthread -lm $(FFLAGS)
+    endif
+  else ifeq ($(FC), gfortran)
+    ifndef LFLAGS
+      LFLAGS = -llapack $(FFLAGS)
+    else
+      LFLAGS += -llapack $(FFLAGS)
+    endif
+  endif
+endif
+
+#get the directory of this makefile
+AMRLIB:=$(CLAW)/amrclaw/src/2d
+GEOLIB:=$(CLAW)/geoclaw/src/2d/shallow
+
+
+#list of common sources for amr 2d codes
+COMMON_MODULES += \
+  $(AMRLIB)/amr_module.f90 \
+  $(AMRLIB)/regions_module.f90 
+
+#list of common sources needed for geoclaw codes (subset of amr dependencies)
+COMMON_MODULES += \
+  $(GEOLIB)/utility_module.f90 \
+  $(GEOLIB)/geoclaw_module.f90 \
+  $(GEOLIB)/surge/holland_storm_module.f90 \
+  $(GEOLIB)/surge/stommel_storm_module.f90 \
+  $(GEOLIB)/surge/constant_storm_module.f90 \
+  $(GEOLIB)/surge/storm_module.f90 \
+  $(GEOLIB)/multilayer/multilayer_module.f90 \
+  $(GEOLIB)/multilayer/gauges_module.f90 \
+  $(GEOLIB)/topo_module.f90 \
+  $(GEOLIB)/qinit_module.f90 \
+  $(GEOLIB)/refinement_module.f90 \
+  $(GEOLIB)/fixedgrids_module.f90 \
+  $(GEOLIB)/fgmax_module.f90 \
+  $(GEOLIB)/friction_module.f90
+
+# list of source files from AMR library. 
+COMMON_SOURCES += \
+  $(AMRLIB)/prefilp.f90 \
+  $(AMRLIB)/trimbd.f90 \
+  $(AMRLIB)/bound.f90 \
+  $(AMRLIB)/intfil.f90 \
+  $(AMRLIB)/flagregions2.f90  \
+  $(AMRLIB)/quick_sort1.f \
+  $(AMRLIB)/estdt.f \
+  $(AMRLIB)/check4nans.f90 \
+  $(AMRLIB)/init_iflags.f \
+  $(AMRLIB)/igetsp.f \
+  $(AMRLIB)/reclam.f \
+  $(AMRLIB)/birect.f \
+  $(AMRLIB)/cleanup.f \
+  $(AMRLIB)/colate2.f \
+  $(AMRLIB)/errest.f \
+  $(AMRLIB)/prepregstep.f \
+  $(AMRLIB)/prepbigstep.f \
+  $(AMRLIB)/bufnst2.f \
+  $(AMRLIB)/flagger.f \
+  $(AMRLIB)/errf1.f \
+  $(AMRLIB)/fixcapaq.f \
+  $(AMRLIB)/flglvl2.f \
+  $(AMRLIB)/fluxad.f \
+  $(AMRLIB)/fluxsv.f \
+  $(AMRLIB)/grdfit2.f \
+  $(AMRLIB)/moment.f \
+  $(AMRLIB)/nestck2.f \
+  $(AMRLIB)/prepc.f \
+  $(AMRLIB)/prepf.f \
+  $(AMRLIB)/projec2.f \
+  $(AMRLIB)/signs.f \
+  $(AMRLIB)/findcut.f \
+  $(AMRLIB)/smartbis.f \
+  $(AMRLIB)/putnod.f \
+  $(AMRLIB)/putsp.f \
+  $(AMRLIB)/regrid.f \
+  $(AMRLIB)/setuse.f \
+  $(AMRLIB)/stst1.f \
+  $(AMRLIB)/nodget.f \
+  $(AMRLIB)/basic.f \
+  $(AMRLIB)/outval.f \
+  $(AMRLIB)/copysol.f \
+  $(AMRLIB)/outvar.f \
+  $(AMRLIB)/outmsh.f \
+  $(AMRLIB)/outtre.f \
+  $(AMRLIB)/domain.f  \
+  $(AMRLIB)/cellave.f \
+  $(AMRLIB)/fdisc.f \
+  $(AMRLIB)/fss.f \
+  $(AMRLIB)/zeroin.f \
+  $(AMRLIB)/setflags.f \
+  $(AMRLIB)/shiftset2.f \
+  $(AMRLIB)/conck.f \
+  $(AMRLIB)/domshrink.f \
+  $(AMRLIB)/domprep.f \
+  $(AMRLIB)/domup.f \
+  $(AMRLIB)/domcopy.f \
+  $(AMRLIB)/setdomflags.f \
+  $(AMRLIB)/setIndices.f \
+  $(AMRLIB)/coarseGridFlagSet.f \
+  $(AMRLIB)/addflags.f \
+  $(AMRLIB)/baseCheck.f \
+  $(AMRLIB)/drivesort.f \
+  $(AMRLIB)/flagcheck.f \
+  $(AMRLIB)/domgrid.f \
+  $(AMRLIB)/setPhysBndryFlags.f \
+  $(AMRLIB)/griddomup.f \
+  $(AMRLIB)/griddomcopy.f \
+  $(AMRLIB)/griddomshrink.f \
+  $(AMRLIB)/coarsen.f \
+  $(AMRLIB)/auxcoarsen.f \
+  $(AMRLIB)/intcopy.f \
+  $(AMRLIB)/preintcopy.f \
+  $(AMRLIB)/icall.f \
+  $(AMRLIB)/preicall.f \
+  $(AMRLIB)/inlinelimiter.f \
+  $(AMRLIB)/cstore.f \
+  $(AMRLIB)/saveqc.f \
+  $(AMRLIB)/opendatafile.f \
+  $(AMRLIB)/init_bndryList.f \
+  $(AMRLIB)/resize_bndryList.f \
+  $(AMRLIB)/init_nodes.f90 \
+  $(AMRLIB)/restrt_nodes.f90 \
+  $(AMRLIB)/resize_nodes.f90 \
+  $(AMRLIB)/init_alloc.f90 \
+  $(AMRLIB)/restrt_alloc.f90 \
+  $(AMRLIB)/resize_alloc.f90
+
+# list of source files from GEOCLAW library. 
+COMMON_SOURCES += \
+  $(GEOLIB)/setprob.f90 \
+  $(GEOLIB)/qinit.f90 \
+  $(GEOLIB)/topo_update.f90 \
+  $(GEOLIB)/cellgridintegrate2.f \
+  $(GEOLIB)/topointegral.f \
+  $(GEOLIB)/bilinearintegral.f \
+  $(GEOLIB)/src1d.f90 \
+  $(GEOLIB)/flux2fw.f \
+  $(GEOLIB)/qad.f \
+  $(GEOLIB)/bc2amr.f \
+  $(GEOLIB)/allowflag.f90  \
+  $(GEOLIB)/tick.f \
+  $(GEOLIB)/setgrd.f \
+  $(GEOLIB)/gfixup.f \
+  $(GEOLIB)/ginit.f \
+  $(GEOLIB)/advanc.f \
+  $(GEOLIB)/amr2.f90 \
+  $(GEOLIB)/fgmax_read.f90 \
+  $(GEOLIB)/fgmax_frompatch.f90 \
+  $(GEOLIB)/fgmax_interpolate0.f90 \
+  $(GEOLIB)/fgmax_values.f90 \
+  $(GEOLIB)/fgmax_finalize.f90 \
+  $(GEOLIB)/check.f \
+  $(GEOLIB)/restrt.f \
+  $(GEOLIB)/multilayer/update.f90 \
+  $(GEOLIB)/multilayer/setaux.f90 \
+  $(GEOLIB)/multilayer/flag2refine2.f90  \
+  $(GEOLIB)/multilayer/getmaxspeed.f90 \
+  $(GEOLIB)/multilayer/b4step2.f90 \
+  $(GEOLIB)/multilayer/upbnd.f  \
+  $(GEOLIB)/multilayer/valout.f90 \
+  $(GEOLIB)/multilayer/filval.f90 \
+  $(GEOLIB)/multilayer/filpatch.f90 \
+  $(GEOLIB)/multilayer/step2.f90 \
+  $(GEOLIB)/multilayer/stepgrid.f \
+  $(GEOLIB)/multilayer/src2.f90