--- a/src/2d/shallow/valout.f90
+++ b/src/2d/shallow/valout.f90
@@ -27,6 +27,7 @@ subroutine valout(level_begin, level_end, time, num_eqn, num_aux)
     real(kind=8), intent(in) :: time
 
     ! Locals
+    logical :: timing_file_exists
     integer, parameter :: out_unit = 50
     integer :: i, j, m, level, output_aux_num, num_stop, digit, num_dim
     integer :: grid_ptr, num_cells(2), num_grids, q_loc, aux_loc
@@ -401,8 +402,8 @@ subroutine valout(level_begin, level_end, time, num_eqn, num_aux)
     ! ==========================================================================
     ! Write out timing stats
     ! Assume that this has been started some where
-    ! if (.not.exists(timing_file_name)) then
-    if (frame == 0) then
+    inquire(file=timing_file_name, exist=timing_file_exists)
+    if (.not. timing_file_exists) then
         ! Write header out and continue
         open(unit=out_unit, file=timing_file_name, form='formatted',         &
              status='unknown', action='write')