--- a/src/2d/shallow/valout.f90
+++ b/src/2d/shallow/valout.f90
@@ -380,25 +380,8 @@ subroutine valout(level_begin, level_end, time, num_eqn, num_aux)
 
     ! ==========================================================================
     ! Write out timing stats
-    ! Create a new file if frame is set to 0 or if it does not exist yet
-    ! Note that this does not currently really handle restarts.
-    inquire(file=timing_file_name, exist=timing_file_exists)
-    if (frame == 0 .or. (.not.timing_file_exists)) then
-        ! Write header out and continue
-        open(unit=out_unit, file=timing_file_name, form='formatted',         &
-             status='unknown', action='write')
-        ! Construct header string
-        timing_line = 'output_time,total_wall_time,total_cpu_time,'
-        timing_substr = ""
-        do level=1, mxnest
-            write(timing_substr, timing_header_format) level, level, level
-            timing_line = trim(timing_line) // trim(timing_substr)
-        end do
-        write(out_unit, "(a)") timing_line
-    else
-        open(unit=out_unit, file=timing_file_name, form='formatted',         &
+    open(unit=out_unit, file=timing_file_name, form='formatted',         &
              status='old', action='write', position='append')
-    end if
     
     timing_line = "(e16.6, ', ', e16.6, ', ', e16.6,"
     do level=1, mxnest