Following up on the changes introduced in https://github.com/clawpack/amrclaw/pull/202 this introduces similar changes geared towards GeoClaw.