@mandli 2018-08-01
Not particular important but I did look at timing stats for a couple of runs:

# Timing Stats

## Debug

### New valout
```
Integration Time (stepgrid + BC + overhead)
Level           Wall Time (seconds)    CPU Time (seconds)   Total Cell Updates
  1                     0.520                  0.522            0.864E+05
  2                     2.093                  2.094            0.358E+06
  3                    95.063                 94.989            0.195E+08
total                  97.676                 97.605            0.200E+08

All levels:
stepgrid               93.936                 93.879
BC/ghost cells          3.713                  3.694
Regridding              0.579                  0.579
Output (valout)         1.931                  1.912

Total time:           100.356                100.267
Using  1 thread(s)

Note: The CPU times are summed over all threads.
      Total time includes more than the subroutines listed above
```
### Old Valout

```
Integration Time (stepgrid + BC + overhead)
Level           Wall Time (seconds)    CPU Time (seconds)   Total Cell Updates
  1                     0.524                  0.525            0.864E+05
  2                     2.114                  2.119            0.358E+06
  3                    95.404                 95.347            0.195E+08
total                  98.042                 97.991            0.200E+08

All levels:
stepgrid               94.342                 94.282
BC/ghost cells          3.668                  3.677
Regridding              0.572                  0.571
Output (valout)         2.103                  0.000

Total time:           100.881                100.809
Using  1 thread(s)

Note: The CPU times are summed over all threads.
      Total time includes more than the subroutines listed above

```

## Optimized

### New Valout

```
Integration Time (stepgrid + BC + overhead)
Level           Wall Time (seconds)    CPU Time (seconds)   Total Cell Updates
  1                     0.178                  0.178            0.864E+05
  2                     0.601                  0.723            0.358E+06
  3                     7.726                 29.036            0.195E+08
total                   8.505                 29.937            0.200E+08

All levels:
stepgrid                7.602                 27.425
BC/ghost cells          0.880                  2.491
Regridding              0.182                  0.340
Output (valout)         1.847                  1.830

Total time:            10.557                 32.159
Using  4 thread(s)

Note: The CPU times are summed over all threads.
      Total time includes more than the subroutines listed above
```

### Old Valout

```
Integration Time (stepgrid + BC + overhead)
Level           Wall Time (seconds)    CPU Time (seconds)   Total Cell Updates
  1                     0.182                  0.192            0.864E+05
  2                     0.622                  0.737            0.358E+06
  3                     7.845                 29.374            0.195E+08
total                   8.649                 30.303            0.200E+08

All levels:
stepgrid                7.716                 27.709
BC/ghost cells          0.915                  2.572
Regridding              0.186                  0.347
Output (valout)         2.081                  0.000

Total time:            10.939                 32.724
Using  4 thread(s)

Note: The CPU times are summed over all threads.
      Total time includes more than the subroutines listed above
```