@mandli 2018-08-02
I think I fixed it.  This change also may make restarts work as well.