@rjleveque 2018-08-02
@mandli: The `_output/timing.csv` file isn't getting overwritten properly if you do a new run that is not a restart.  It appends the new data to the end, so it goes from time 0 up to tfinal twice over.