@mandli 2018-08-03
I now added a bit more to the logic:
```fortran
if (frame == 0 .or. (.not.timing_file_exists)) then
```
This way if the file does not exist or the frame is 0 it creates/overwrites the existing file.