@mandli 2018-08-03
There's a couple of other bugs I found so don't merge this in quite yet.  I also need to convert the multilayer `valout` routine.