@mandli 2018-08-04
I think we have everything in this commit at this point.  I have tested binary output in both the single and two-layer cases as well and everything seems to be working.  One thing I realized though is that we do not have a check for the binary output.  Something to raise an issue here and in clawpack/amrclaw.