@mandli 2018-08-07
This has now incorporated the suggestion in clawpack/amrclaw#226 put the question of where to put the timing data in a file rather than just the console still needs to be resolved for GeoClaw.